requirements.txt
tweepy
attrs
requests
